Server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
pyjwt
passlib[bcrypt]
bcrypt==4.0.1